+++ src/System.Buffers.Experimental/System/Buffers/OwnedMemory.cs
@@ -1,27 +1,11 @@
-﻿using System.Runtime.CompilerServices;
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
 namespace System.Buffers
 {
-    public sealed class OwnedArray<T> : OwnedMemory<T>
-    {
-        public new T[] Array => base.Array;
-
-        public static implicit operator T[](OwnedArray<T> owner) {
-            return owner.Array;
-        }
-
-        public static implicit operator OwnedArray<T>(T[] array) {
-            return new OwnedArray<T>(array);
-        }
-
-        public OwnedArray(int length) : base(new T[length], 0, length)
-        {}
-
-        public OwnedArray(T[] array) : base(array, 0, array.Length)
-        {}
-    }
-
     public class OwnedNativeMemory : OwnedMemory<byte>
     {
         public OwnedNativeMemory(int length) : this(length, Marshal.AllocHGlobal(length))
@@ -102,15 +86,5 @@ namespace System.Buffers
             base.Dispose(disposing);
         }
     }
-
-    internal class OwnerEmptyMemory<T> : OwnedMemory<T>
-    {
-        public readonly static OwnedMemory<T> Shared = new OwnerEmptyMemory<T>();
-        readonly static T[] s_empty = new T[0];
-
-        public OwnerEmptyMemory() : base(s_empty, 0, 0) { }
-
-        protected override void Dispose(bool disposing)
-        {}
-    }
 }
+
