+++ src/System.Buffers.Experimental/project.json
@@ -26,7 +26,8 @@
     "System.Diagnostics.Debug": "4.0.11",
     "System.Slices": { "target": "project" },
     "System.Text.Primitives": { "target": "project" },
-    "System.Threading": "4.0.11"
+    "System.Threading": "4.0.11",
+    "System.Runtime.InteropServices": "4.1.0"
   },
   "frameworks": {
     "netstandard1.1": {
