+++ src/System.Text.Encodings.Web.Utf8/Properties/AssemblyInfo.cs
@@ -14,6 +14,3 @@ using System.Runtime.InteropServices;
 // to COM components.  If you need to access a type in this assembly from
 // COM, set the ComVisible attribute to true on that type.
 [assembly: ComVisible(false)]
-
-// The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("20974d06-ebdc-44b8-a034-de1440cf7a45")]
