+++ src/System.Text.Utf8/project.json
@@ -12,7 +12,7 @@
     ],
     "releaseNotes": "Pre-release package, for testing only",
     "licenseUrl": "http://go.microsoft.com/fwlink/?LinkId=329770",
-    "iconUrl": "https://go.microsoft.com/fwlink/?linkid=833199",
+    "iconUrl": "http://go.microsoft.com/fwlink/?linkid=833199",
     "projectUrl": "https://github.com/dotnet/corefxlab",
     "requireLicenseAcceptance": true
   },
